requirements.txt
aenum==3.1.15
aiohttp==3.9.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
blinker==1.7.0
branca==0.7.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.7
django-allauth==0.58.2
django-crispy-forms==2.1
django-cron==0.6.0
django-crontab==0.7.1
django-mapbox-location-field==2.0.0
django-plotly-dash==2.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
dpd-components==0.1.0
et-xmlfile==1.1.0
Flask==3.0.2
folium==0.15.1
frozenlist==1.4.0
future==0.18.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.5
line-bot-sdk==3.5.1
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.4
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
Pillow==10.1.0
plotly==5.19.0
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
XlsxWriter==3.2.0
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
